refactor(core): move browserContext and docusaurusContext out of client exports
