+++ packages/docusaurus-theme-common/src/utils/__tests__/useAlternatePageUtils.test.tsx
@@ -9,7 +9,7 @@ import React from 'react';
 import {useAlternatePageUtils} from '../useAlternatePageUtils';
 import {renderHook} from '@testing-library/react-hooks';
 import {StaticRouter} from 'react-router-dom';
-import {Context} from '@docusaurus/docusaurusContext';
+import {Context} from '@docusaurus/core/src/client/docusaurusContext';
 import type {DocusaurusContext} from '@docusaurus/types';
 
 describe('useAlternatePageUtils', () => {
