+++ packages/docusaurus-theme-common/src/utils/__tests__/usePluralForm.test.tsx
@@ -9,7 +9,7 @@ import {jest} from '@jest/globals';
 import React from 'react';
 import {usePluralForm} from '../usePluralForm';
 import {renderHook} from '@testing-library/react-hooks';
-import {Context} from '@docusaurus/docusaurusContext';
+import {Context} from '@docusaurus/core/src/client/docusaurusContext';
 import type {DocusaurusContext} from '@docusaurus/types';
 
 describe('usePluralForm', () => {
