+++ packages/docusaurus/src/client/App.tsx
@@ -9,8 +9,8 @@ import React from 'react';
 
 import routes from '@generated/routes';
 import renderRoutes from './exports/renderRoutes';
-import {BrowserContextProvider} from './exports/browserContext';
-import {DocusaurusContextProvider} from './exports/docusaurusContext';
+import {BrowserContextProvider} from './browserContext';
+import {DocusaurusContextProvider} from './docusaurusContext';
 import PendingNavigation from './PendingNavigation';
 import BaseUrlIssueBanner from './baseUrlIssueBanner/BaseUrlIssueBanner';
 import SiteMetadataDefaults from './SiteMetadataDefaults';
