+++ packages/docusaurus/src/client/exports/__tests__/BrowserOnly.test.tsx
@@ -12,7 +12,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 import BrowserOnly from '../BrowserOnly';
-import {Context} from '../browserContext';
+import {Context} from '../../browserContext';
 
 describe('<BrowserOnly>', () => {
   const originalEnv = process.env;
